+++ shuup/core/models/_orders.py
@@ -299,10 +299,10 @@ class Order(MoneyPropped, models.Model):
         taxful_total = TaxfulPrice(0, self.currency)
         taxless_total = TaxlessPrice(0, self.currency)
         for line in self.lines.all():
-            taxful_total += line.taxful_price
-            taxless_total += line.taxless_price
-        self.taxful_total_price = _round_price(taxful_total)
-        self.taxless_total_price = _round_price(taxless_total)
+            taxful_total += _round_price(line.taxful_price)
+            taxless_total += _round_price(line.taxless_price)
+        self.taxful_total_price = taxful_total
+        self.taxless_total_price = taxless_total
 
     def _cache_contact_values(self):
         sources = [